added news, created news layout, rm twitter feed
Since website was created, the page for the full list of News was badly
formatted. The nice format only worked for the post type of page, but
that was limited to the scicomm posts. Here I duplicated the posts.html
inside the layout folder of the theme, and renamed it to news.html. That
gave the nice formatting! It's possible that in future updates of the
theme, this file can be removed automatically, I can create again or
put it inside a layout folder in my website root.
Also, on the current online version the iframes are not working, I don't
know why. Couldn't find people complaining about this, so for now I
removed the widget for the twitter feed. iframes for blog posts and
google maps on the home page are still not showing, but will leave the
code as is for now (hopefully will be fixed by itself if I need to
update R/package again in the future).